requirements.txt
pandas
spacy
sentence-transformers
faiss-cpu
streamlit
